move appropriate unit tests from engine/test to netlists/test. Not all tests passing yet, but close
